Testing h5

+++ gene_transformer/dataset.py
@@ -138,7 +138,7 @@ class H5Dataset(Dataset):
             # Squeeze so that batched tensors end up with (batch_size, seq_length)
             # instead of (batch_size, 1, seq_length)
             fields["input_ids"].append(batch_encoding["input_ids"].astype(np.int8))
-            fields["attention_mask"].append(batch_encoding["attention_mask"])
+            fields["attention_mask"].append(batch_encoding["attention_mask"].astype(np.int8))
             fields["id"].append(np.array(seq_record.id, dtype=object))
             fields["description"].append(np.array(seq_record.description, dtype=object))
             fields["sequence"].append(np.array(str(seq_record.seq).upper(), dtype=object))
